Refactor: 장바구니 기능 수정 #15

+++ src/main/java/com/hanhwa_tae/gulhan/cart/command/domain/repository/CartRepository.java
@@ -15,4 +15,4 @@ public interface CartRepository{
 
     @Modifying
     @Query("DELETE FROM Cart c WHERE c.userNo.userNo = :userNo")
-    void deleteAllByUserNo(@Param("userNo") int userNo);}
+    void deleteAllByUserNo(@Param("userNo") Long userNo);}
